requirements.txt
requests
xmltodictsource